requirements.txt
appdirs==1.4.4
apt-xapian-index==0.49
asgiref==3.8.1
attrs==21.2.0
bcc==0.18.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.7
django-bootstrap4==24.3
django-stdimage==6.0.2
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
fonttools==4.29.1
fs==2.4.12
fuse-python==1.0.2
gast==0.5.2
gpg==1.16.0
greenlet==1.1.2
gunicorn==22.0.0
hidpidaemon==18.4.6
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
jsonschema==3.2.0
kernelstub==3.1.4
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Markdown==3.3.6
matplotlib==3.5.1
more-itertools==8.10.0
mpmath==0.0.0
msgpack==1.0.3
mysql==0.0.3
mysqlclient==2.2.4
netaddr==0.8.0
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
packaging==21.3
Pillow==9.0.1
ply==3.11
pop-transition==1.1.2
protobuf==3.12.4
PyAudio==0.2.11
pycairo==1.20.1
pycups==2.0.1
pydbus==0.6.0
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pylibacl==0.6.0
pymacaroons==0.13.0
PyMySQL==1.1.1
PyNaCl==1.5.0
pynvim==0.4.2
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQtWebEngine==5.15.5
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-gnupg==0.4.8
python-xlib==0.29
pythran==0.10.0
pytz==2022.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
repolib==2.2.1
repoman==1.4.0
requests==2.25.1
scipy==1.8.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
Send2Trash==1.8.1b0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
sympy==1.9
systemd-python==234
texttable==1.6.4
tornado==6.1
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unicodedata2==14.0.0
urllib3==1.26.5
vboxapi==1.0
wadllib==1.3.6
webencodings==0.5.1
websocket-client==1.2.3
whitenoise==6.7.0
xdg==5
xkit==0.0.0
zipp==1.0.0
